initial.rev26
-made so detects if scrollbars vis on PUIcs, if its NOT vis then it wont
hide them and do nothing on collapse, if it IS vis then it will hide
them then reshow on collapse
-:todo:it looks like expandedHeight should
be cumHeight - 1 but i didnt implement this
-:todo:watch out for content
of stack showing someday, right now even with not specifying
overflow:hidden the contents of stack doesnt show if the height is set
to something less than max height
-:bug: possible bug to report to
bugzilla: if panel is visible and if stack is empty and stack is set to
height of 38px, its weird but the box comes out to be 50px;. no matter
what if stack empty set stack to whatever height, the box is always
stack height + 12 UNTIL you put a child into stack, weird
-:learned:
when trying to put a ProflistLoading id'ed toolbarbutton into stack it
woundnt get added when first letter was caps
-tested to see how it works
when os.file lags, by using settimeout and what if stack was expanded,
it is modified approparately now
-fixed: bug, to see if the stackJSON
value was in pathsInIni it was doing m.props.profName in
profToolkit.pathsInIni, but pathsInIni is an array so I have to use
indexOf
-i accidentally had a typo in the collapse function, there was a
z in front of the line that set the height back to collapsedHeight, this
just happend today while i was working, was not there in prior
versions
-:learned: if add dynamically createElement and then add it to
a panel thats hidden, and while its hidden (meaning state has never been
showing or open since you added the element), if you do boxObject.height
or window.getComputedStyle(el).getPropertyValue('height') they will
return 0. this is true even if you set el.style.height to say '38px'
after adding the element, you can read this '38px' though by going
el.style.height.
-:learned: if add dynamically createElement and then
add it to a panel thats hidden, and then open the panel (in my testing i
ran getComputed and boxObject on it, but i think the following is true
even if you dont run getComputed or boxObject on the element while the
panel was showing/open), and then close the panel, then if you go
boxObject.height or
window.getComputedStyle(el).getPropertyValue('height') it will return to
you the LAST KNEW, meaning the height it was when the panel was last
open. so after its closed even if you set el.style.height to '38px' and
run boxObject or getComputedStyle it will still return to you the last
known height (again: last known is when it was last open), you can get
this string value of style.height though by going
el.style.height
-:learned: boxObject always returns unitless so its just
a number but the units are always px
-:learned: getComputedStyle always
returns with units and i suspect its the units used by the css style
applied (if no css style changed the height then it will default to 'px'
i suspect)
-:learned: boxObject is rounded value while getComputedStyle
gives float value like if el.boxObject.height gives you 38px
window.getComputedStyle(el).getPropertyValue('height') will be anything
from 37.5px-38.4px
-:todo: im thinking because right now (as i dont have
access to file watcher) im always refreshingIni on show of popup, if
after running create/delete/rename profile promise complete if the panel
is open then refresh, if its not open then dont refresh
-fixed: in
`updateStackDOMJson_basedOnToolkit` it was splicing in every time, i
corrected the if, im pretty sure its a bug fix, so it only splices if
pos of path was -1 in stackJSON, i think this fixes the duplicate
triplicate bug i just saw within the last hour of 1:37am
-fixed: when
renaming, in revamped method, if open menu, its still reading file, and
before it finishes i go to rename and rename a profile, then
toolbarbutton is blank in all windows for this renamed profile
-chagned
so identifier is path in ini for toolbarbuttons for profile buttons and
just the label name for non profile buttons
-partially revamped
`updateStackDOMJson_basedOnToolkit` and `updateMenuDOM`, now it always
runs updateMenuDOM and figs if it needs a dom update, if it does and
find dont update stack is true then it reports so.
-also fixed if open
new window it will properly update the menu