1.2.rev62
### bootstrap.js
* properly adding addEventListner to the toolbarbuttons
* before i was checking if it had status attribute on toolbarbutton
before testing if it should be removed. now if its not a profile button
(indicated by not having a status attribute) then if it doesnt have the
`profilist-do_not_auto_remove` class it will be removed
* if profilist_box is set to display flex than the stack element within
does not work. so i removed the display:flex which i had introduced in
order to fix that 1px on top of the stack but i removed it and it doesnt
seem to hvae that 1px on top bug anymore
* i know i was force setting el.style.height on the box's but i stopped
because for some weird reason sometimes it finds that profilist-loading
height is 37 instead of 38 sometimes (i think its getting 37 because its
not getting the styling from tbb xbl as its load is not as fast as the
calc is happening before? maybe?)
* found another fix for this 1px on top thing. it also fixes heights of
tbb-box and tbb. had to set moz-image-region to 16, 16
* by setting border-box and it has now made tbb have proper height, so
we now get proper PUIsync_height of 38
* by removing `display="xul:button"` from xbl of tbb `command` event no
longer works, so switched to using `click`