1.2.rev65
## bootstrap.js
* when clicked on toolbarbutton-icon of create profile it wasnt doing
anything, it should redir to create new profile, it does this now

## xbl.xml
* the weirdest weirdest fix for width of panel not growing because of
width of submenu was to use `max-width:0;` on `profilist-tbb-box` so
weird i still dont understand it and cant explain it
* introduced html namespace prep for input element

##main.css
* main this is now i support ellipsis when label is too long and submenu
expands
* switched from left to width, was doing left because i couldnt figure
out how to get width to work it would keep squishing, i fixed this by
giving sub-elements flex-shrink:0
* *learned* if want to do a 0ms duration must type in `0ms` not `0`
otherwise it messes up and doesnt read the delay time right. it will
read the delay as default which is 0ms. it probably wont read the
transition-function param either. `transition: width 0ms ease 2200ms`