1.2.rev212
COMMIT TO SAVE
tested though at this point so have landed do_sysLaunchers_Launcher
support in updateIconTo_SysLaunchers_Launcher_RunningWins_ofExePath.

So this effort is for when in CP user changes base_icon find all
launchers and running windows with path of exe_path, and update its icon
to be new_icon_path.