3.0a.rev56
* Ok the mac launcher is being created with symlink method, however that
method is the one that breaks the update mechanism. I couldnt get the
symlink method to work, it would always say the launcher .app is corrupt
whenever i tried to launch it. I need to try to get that to work, then
mayyybe try hard link, but hopefully can get alias's to work.