3.0a.rev58
* Commit to save somewhat, but I tested installing the addon and there
are no errors, so half tested

* Changed getRunningExePathForProfFromPlat to
getLastExePathForProfFromFS as I remembered compat ini has that info,
which is awesome ah - i wrote up this function as well but UNTESTED
* I changed up the other half of getRunningExePathForProfFromPlat to
getIsRunningFromIni, which uses the os.file.open and fcntl methods. I
also wrote that up but UNTESTED

* added HollowStructure to cutils.jsm

## :TODO:
* Now that I seperated the above into two functions, I shouldn't need
exePath specifified if the profile is running, so I need to go in and
adjust that, as i am now voiding that assumption. exePath should only be
needed if "dev mode is on && the profile is running" ---- figure out if
i need to exePath for the currently running profile for dom/bl layers if
dev mode is off