3.0a.rev88
CSS fun bordered the whole panel, but the fun part was bordering the
arrow with drop shadow. And the cool thing i didnt mention about earlier
is there is no overlap of dropshadow by the arrow on the contents
dropshadow as a the arrow was completed by the before and after pseudo
elements. saved a graphic that shows how the bordering was done with
different elements. the yellow covers the drop shadow of the after
element.