3.0b4.rev14
* Fixed huge issue, in that the linux/mac testing for lock was doing
ostypes.API('open') with O_CREAT which is a big no. For some reason
though, even though I would call ostypes.API('close') on it, it would
not close it and then later on it was an inaccessible file by my lock
checker.
* Updated SIPWorker from NativeShot bootstrap and MainWorker
* There were some spots in bootstrap that didnt have an array for 2nd
arg of .post so i have no idea how it was working, but added that in
* ostypes_x11 had some updates